refactor(backups): consolidate recovery key and discovery into engine and backup files
