--- backend/internal/backup/recovery_key.go
@@ -1,118 +1,0 @@
-package backup
-
-import (
-	"context"
-	"crypto/rand"
-	"encoding/base32"
-	"fmt"
-	"regexp"
-	"strings"
-
-	"emperror.dev/errors"
-	"go.getarcane.app/sys/crypto"
-	"gorm.io/gorm"
-
-	"github.com/getarcaneapp/arcane/backend/v2/internal/database"
-)
-
-// RecoveryKeyConfigID is the singleton row holding the instance-wide backup
-// recovery key.
-const RecoveryKeyConfigID = "system-recovery"
-
-// ErrRecoveryKeyNotConfigured reports that no recovery key has been stored yet.
-var ErrRecoveryKeyNotConfigured = errors.New("recovery key is not configured")
-
-// RecoveryKeyFormat is the only accepted key shape, generated by
-// GenerateRecoveryKey: 48 base32 characters (~240 bits from crypto/rand) in 8
-// hyphenated groups for hand transcription. The key is the Rustic repository
-// password for system recovery and volume backup repositories and is used
-// exactly as displayed, hyphens included.
-var RecoveryKeyFormat = regexp.MustCompile(`^[A-Z2-7]{6}(-[A-Z2-7]{6}){7}$`)
-
-// SystemBackupRecoveryConfig stores the encrypted recovery key. The key must be
-// reversible because it is passed to Rustic as the repository password.
-type SystemBackupRecoveryConfig struct {
-	database.BaseModel
-
-	EncryptedRecoveryKey string `gorm:"column:encrypted_recovery_key;type:text;not null"`
-}
-
-func (SystemBackupRecoveryConfig) TableName() string { return "system_backup_recovery_config" }
-
-// ValidateRecoveryKey rejects keys outside the generated shape.
-func ValidateRecoveryKey(key string) error {
-	if !RecoveryKeyFormat.MatchString(strings.TrimSpace(key)) {
-		return errors.New("enter the generated recovery key (8 groups of 6 characters)")
-	}
-	return nil
-}
-
-// GenerateRecoveryKey returns a fresh 240-bit key in the display format.
-func GenerateRecoveryKey() (string, error) {
-	raw := make([]byte, 30)
-	if _, err := rand.Read(raw); err != nil {
-		return "", fmt.Errorf("failed to generate recovery key: %w", err)
-	}
-	encoded := base32.StdEncoding.WithPadding(base32.NoPadding).EncodeToString(raw)
-	groups := make([]string, 0, len(encoded)/6)
-	for i := 0; i < len(encoded); i += 6 {
-		groups = append(groups, encoded[i:i+6])
-	}
-	return strings.Join(groups, "-"), nil
-}
-
-// RecoveryKeyStore persists the shared backup recovery key. Both the system
-// backup and volume backup domains resolve their Rustic repository password
-// through it, so it lives outside either domain.
-type RecoveryKeyStore struct {
-	db *database.DB
-}
-
-func NewRecoveryKeyStore(db *database.DB) *RecoveryKeyStore {
-	return &RecoveryKeyStore{db: db}
-}
-
-// Configured reports whether a recovery key has been stored.
-func (s *RecoveryKeyStore) Configured(ctx context.Context) (bool, error) {
-	var count int64
-	if err := s.db.WithContext(ctx).Model(&SystemBackupRecoveryConfig{}).
-		Where("id = ? AND encrypted_recovery_key <> ''", RecoveryKeyConfigID).Count(&count).Error; err != nil {
-		return false, fmt.Errorf("failed to load recovery key status: %w", err)
-	}
-	return count > 0, nil
-}
-
-// Get returns the stored recovery key, decrypting it at rest. Returns
-// ErrRecoveryKeyNotConfigured when no key has been stored.
-func (s *RecoveryKeyStore) Get(ctx context.Context) (string, error) {
-	var config SystemBackupRecoveryConfig
-	if err := s.db.WithContext(ctx).
-		Where("id = ? AND encrypted_recovery_key <> ''", RecoveryKeyConfigID).
-		First(&config).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return "", ErrRecoveryKeyNotConfigured
-		}
-		return "", fmt.Errorf("failed to load recovery key: %w", err)
-	}
-	key, err := crypto.Decrypt(config.EncryptedRecoveryKey)
-	if err != nil {
-		return "", fmt.Errorf("failed to decrypt recovery key: %w", err)
-	}
-	return key, nil
-}
-
-// Set validates and stores the recovery key, replacing any previous value.
-func (s *RecoveryKeyStore) Set(ctx context.Context, recoveryKey string) error {
-	if err := ValidateRecoveryKey(recoveryKey); err != nil {
-		return err
-	}
-	encrypted, err := crypto.Encrypt(recoveryKey)
-	if err != nil {
-		return fmt.Errorf("failed to encrypt recovery key: %w", err)
-	}
-	config := SystemBackupRecoveryConfig{ID: RecoveryKeyConfigID, EncryptedRecoveryKey: encrypted}
-	if err := s.db.WithContext(ctx).Save(&config).Error; err != nil {
-		return fmt.Errorf("failed to save recovery key: %w", err)
-	}
-	return nil
-}
